+++ dorado-core/src/main/java/ai/houyi/dorado/rest/http/impl/Webapp.java
@@ -30,9 +30,8 @@ import ai.houyi.dorado.rest.controller.RootController;
 import ai.houyi.dorado.rest.http.Filter;
 import ai.houyi.dorado.rest.http.MethodReturnValueHandler;
 import ai.houyi.dorado.rest.http.MethodReturnValueHandlerConfig;
-import ai.houyi.dorado.rest.router.UriRoutingRegistry;
+import ai.houyi.dorado.rest.router.trie.UriRoutingRegistry;
 import ai.houyi.dorado.rest.router.trie.Router;
-import ai.houyi.dorado.rest.util.LogUtils;
 import ai.houyi.dorado.rest.util.PackageScanner;
 
 /**
